Create InfraredSensor.java
This class should now be in the sensors package along with the other sensors. Changes were made to the class in terms of confidentiality between variables and functions. So, all but two of the get and set functions were removed with a new function named CalculateInfraredReading to initialise infraredReading and angleReading for a sensor object.